runtime: reduce node types controller binding hotspot
