+++ doc/.governance/rust-oversized-file-baseline.tsv
@@ -4,7 +4,6 @@ code	crates/oasis7/src/bin/oasis7_web_launcher/control_plane.rs	1202
 code	crates/oasis7/src/viewer/runtime_live/llm_sidecar.rs	1389
 code	crates/oasis7_client_launcher/src/launcher_core.rs	1254
 code	crates/oasis7_node/src/replication.rs	1221
-code	crates/oasis7_node/src/types.rs	1277
 test	crates/oasis7/src/simulator/llm_agent/tests_split_part1.rs	1204
 test	crates/oasis7/src/viewer/runtime_live/tests/auth_actions.rs	1335
 test	crates/oasis7_net/src/tests.rs	1239
