+++ doc/engineering/rust-governance/rust-1200-line-root-cause-governance-2026-03-29.project.md
@@ -47,6 +47,7 @@
   - `libp2p-net-orchestrator-burn-down`：`crates/oasis7_net/src/libp2p_net.rs` 新增 `constructor_support.rs`，将 `Libp2pNetwork::new` 内重复的 republish / discovery refresh / bootstrap redial / initial bootstrap dials 调度逻辑外提；根文件已从 1199 行降到 1156 行，`cargo test -p oasis7_net --lib -- --nocapture` 122 项通过，`check-rust-file-size` 保持 structural slicing baseline 不变。
   - `rust-node-types-controller-binding-burn-down`：`crates/oasis7_node/src/types.rs` 新增子模块 `types/main_token_controller_binding.rs`，将 main-token controller binding 常量、默认值、builder 与校验 helper 外提，根文件已从 1277 行降到 1069 行，并同步从 `rust-oversized-file-baseline.tsv` 退休旧冻结条目；`cargo check -p oasis7_node`、`cargo test -p oasis7_node types::tests:: -- --nocapture` 以及两条 controller-binding 授权定向测试通过，`cargo test -p oasis7_node --lib -- --nocapture` 仍存在 3 条与 replication/network 相关的现有失败签名，已在 execution log 留痕。
   - `TASK-ENGINEERING-104`：确认 `doc/.governance/rust-oversized-file-baseline.tsv` 在 `2026-03-30` 被提交为只剩注释的空壳，导致 `scripts/check-rust-file-size.sh` 在 required gate 中把全部存量超限项误判为“相对 HEAD 新增”；现已按当前仓库实况重写 baseline，恢复 `scripts/ci-tests.sh required` 的 frozen baseline 语义，并把专项状态改回继续 burn-down。
+  - `rust-node-types-controller-binding-burn-down`：`crates/oasis7_node/src/types.rs` 新增子模块 `types/main_token_controller_binding.rs`，将 main-token controller binding 常量、默认值、builder 与校验 helper 外提，根文件已从 1277 行降到 1069 行，并同步从 `rust-oversized-file-baseline.tsv` 退休旧冻结条目；`cargo check -p oasis7_node`、`cargo test -p oasis7_node types::tests:: -- --nocapture` 以及两条 controller-binding 授权定向测试通过，`cargo test -p oasis7_node --lib -- --nocapture` 仍存在 3 条与 replication/network 相关的现有失败签名，已在 execution log 留痕。
   - `TASK-ENGINEERING-054`：`oasis7_chain_runtime.rs` 拆出 `cli.rs`，`execution_bridge.rs` 迁移为目录模块，门禁基线已退休旧入口超限项。
   - `TASK-ENGINEERING-055`：`viewer/runtime_live.rs` 与测试集拆为目录模块；同时把 `runtime/events.rs`、`state.rs`、`apply_domain_event_*`、`world/event_processing.rs` 压回 1200 行内，清零本批新增 runtime 超限。
   - `TASK-ENGINEERING-056`：`oasis7_viewer` 首批治理与本轮 runtime / launcher 收尾已完成，退役了 `egui_right_panel_player_guide.rs`、`web_test_api.rs`、`viewer_automation.rs` 以及 `oasis7_provider_parity_bench.rs`、`oasis7_pure_api_client.rs`、`oasis7_provider_local_bridge.rs`、`oasis7_web_launcher/control_plane.rs`、`runtime/world/persistence.rs`、`viewer/live_split_part1.rs`、`runtime/world/governance.rs`、`runtime/world/module_actions_impl_part2.rs` 的生产超限基线。
